moved std types to `_std.pyi` file
