+++ python/ry/ryo3/__init__.pyi
@@ -4,13 +4,12 @@ from __future__ import annotations
 
 import datetime as pydt
 import typing as t
-from os import PathLike
-from pathlib import Path
 
 from ry import dirs as dirs  # noqa: RUF100
 from ry import http as http  # noqa: RUF100
 from ry import xxhash as xxhash  # noqa: RUF100
 from ry._types import Buffer as Buffer  # noqa: RUF100
+from ry._types import FsPathLike
 from ry.http import Headers as Headers  # noqa: RUF100
 from ry.http import HttpStatus as HttpStatus  # noqa: RUF100
 
@@ -27,6 +26,7 @@ from ._flate2 import gzip_decode as gzip_decode
 from ._flate2 import gzip_encode as gzip_encode
 from ._fnv import FnvHasher as FnvHasher
 from ._fnv import fnv1a as fnv1a
+from ._fspath import FsPath as FsPath
 from ._globset import Glob as Glob
 from ._globset import GlobSet as GlobSet
 from ._globset import Globster as Globster
@@ -71,6 +71,7 @@ from ._jiter import json_cache_clear as json_cache_clear
 from ._jiter import json_cache_usage as json_cache_usage
 from ._jiter import parse_json as parse_json
 from ._jiter import parse_json_bytes as parse_json_bytes
+from ._quick_maths import quick_maths as quick_maths
 from ._regex import Regex as Regex
 from ._reqwest import HttpClient as HttpClient
 from ._reqwest import ReqwestError as ReqwestError
@@ -85,6 +86,32 @@ from ._size import parse_size as parse_size
 from ._sqlformat import SqlfmtQueryParams as SqlfmtQueryParams
 from ._sqlformat import sqlfmt as sqlfmt
 from ._sqlformat import sqlfmt_params as sqlfmt_params
+from ._std import Duration as Duration
+from ._std import FileType as FileType
+from ._std import Instant as Instant
+from ._std import Metadata as Metadata
+from ._std import canonicalize as canonicalize
+from ._std import copy as copy
+from ._std import create_dir as create_dir
+from ._std import create_dir_all as create_dir_all
+from ._std import exists as exists
+from ._std import instant as instant
+from ._std import is_dir as is_dir
+from ._std import is_file as is_file
+from ._std import is_symlink as is_symlink
+from ._std import metadata as metadata
+from ._std import read as read
+from ._std import read_bytes as read_bytes
+from ._std import read_stream as read_stream
+from ._std import read_text as read_text
+from ._std import remove_dir as remove_dir
+from ._std import remove_dir_all as remove_dir_all
+from ._std import remove_file as remove_file
+from ._std import rename as rename
+from ._std import sleep as sleep
+from ._std import write as write
+from ._std import write_bytes as write_bytes
+from ._std import write_text as write_text
 from ._tokio import asleep as asleep
 from ._tokio import copy_async as copy_async
 from ._tokio import create_dir_async as create_dir_async
@@ -110,6 +137,9 @@ from ._zstd import zstd_decode as zstd_decode
 from ._zstd import zstd_encode as zstd_encode
 from .errors import FeatureNotEnabledError as FeatureNotEnabledError
 
+# =============================================================================
+# CONSTANTS
+# =============================================================================
 __version__: str
 __authors__: str
 __build_profile__: str
@@ -118,303 +148,8 @@ __pkg_name__: str
 __description__: str
 
 # =============================================================================
-# TYPE ALIASES
+# SH
 # =============================================================================
-
-# =============================================================================
-# STD::TIME
-# =============================================================================
-class Duration:
-    ZERO: Duration
-    MIN: Duration
-    MAX: Duration
-    NANOSECOND: Duration
-    MICROSECOND: Duration
-    MILLISECOND: Duration
-    SECOND: Duration
-
-    def __init__(self, secs: int = 0, nanos: int = 0) -> None: ...
-    def __eq__(self, other: object) -> bool: ...
-    def __ne__(self, other: object) -> bool: ...
-    def __lt__(self, other: Duration) -> bool: ...
-    def __le__(self, other: Duration) -> bool: ...
-    def __gt__(self, other: Duration) -> bool: ...
-    def __ge__(self, other: Duration) -> bool: ...
-    def __hash__(self) -> int: ...
-    def __richcmp__(self, other: Duration | pydt.timedelta, op: int) -> bool: ...
-    def __str__(self) -> str: ...
-    def abs_diff(self, other: Duration) -> Duration: ...
-    def sleep(self) -> None: ...
-
-    # =========================================================================
-    # PYTHON_CONVERSIONS
-    # =========================================================================
-    @classmethod
-    def from_pytimedelta(cls: type[Duration], td: pydt.timedelta) -> Duration: ...
-    def to_pytimedelta(self) -> pydt.timedelta: ...
-
-    # =========================================================================
-    # PROPERTIES
-    # =========================================================================
-    @property
-    def is_zero(self) -> bool: ...
-    @property
-    def nanos(self) -> int: ...
-    @property
-    def secs(self) -> int: ...
-    @property
-    def days(self) -> int: ...
-    @property
-    def seconds(self) -> int: ...
-    @property
-    def microseconds(self) -> int: ...
-    @property
-    def subsec_micros(self) -> int: ...
-    @property
-    def subsec_millis(self) -> int: ...
-    @property
-    def subsec_nanos(self) -> int: ...
-
-    # =========================================================================
-    # CLASSMETHODS
-    # =========================================================================
-    @classmethod
-    def from_hours(cls, hours: int) -> Duration: ...
-    @classmethod
-    def from_micros(cls, micros: int) -> Duration: ...
-    @classmethod
-    def from_millis(cls, millis: int) -> Duration: ...
-    @classmethod
-    def from_mins(cls, mins: int) -> Duration: ...
-    @classmethod
-    def from_nanos(cls, nanos: int) -> Duration: ...
-    @classmethod
-    def from_secs(cls, secs: int) -> Duration: ...
-    @classmethod
-    def from_secs_f32(cls, secs: float) -> Duration: ...
-    @classmethod
-    def from_secs_f64(cls, secs: float) -> Duration: ...
-    @classmethod
-    def from_days(cls, days: int) -> Duration: ...
-    @classmethod
-    def from_weeks(cls, weeks: int) -> Duration: ...
-    def as_micros(self) -> int: ...
-    def as_millis(self) -> int: ...
-    def as_nanos(self) -> int: ...
-    def as_secs(self) -> int: ...
-    def as_secs_f32(self) -> float: ...
-    def as_secs_f64(self) -> float: ...
-
-    # =========================================================================
-    # NOT IMPLEMENTED
-    # =========================================================================
-    def checked_add(self, other: Duration) -> Duration | None: ...
-    def checked_div(self, other: Duration) -> Duration | None: ...
-    def checked_mul(self, other: Duration) -> Duration | None: ...
-    def checked_sub(self, other: Duration) -> Duration | None: ...
-    def div_duration_f32(self, other: Duration) -> float: ...
-    def div_duration_f64(self, other: Duration) -> float: ...
-    def div_f32(self, other: float) -> Duration: ...
-    def div_f64(self, other: float) -> Duration: ...
-    def mul_f32(self, other: float) -> Duration: ...
-    def mul_f64(self, other: float) -> Duration: ...
-    def saturating_add(self, other: Duration) -> Duration: ...
-    def saturating_mul(self, other: Duration) -> Duration: ...
-    def saturating_sub(self, other: Duration) -> Duration: ...
-
-class Instant:
-    def __init__(self) -> None: ...
-    @classmethod
-    def now(cls) -> Instant: ...
-    def __str__(self) -> str: ...
-    def __repr__(self) -> str: ...
-    def __eq__(self, other: object) -> bool: ...
-    def __ne__(self, other: object) -> bool: ...
-    def __lt__(self, other: Instant) -> bool: ...
-    def __le__(self, other: Instant) -> bool: ...
-    def __gt__(self, other: Instant) -> bool: ...
-    def __ge__(self, other: Instant) -> bool: ...
-    def __hash__(self) -> int: ...
-    def __add__(self, other: Duration) -> Instant: ...
-    def __iadd__(self, other: Duration) -> Instant: ...
-    @t.overload
-    def __sub__(self, other: Duration) -> Instant: ...
-    @t.overload
-    def __sub__(self, other: Instant) -> Duration: ...
-
-    # =========================================================================
-    # INSTANCE METHODS
-    # =========================================================================
-    def checked_add(self, other: Duration) -> Instant | None: ...
-    def checked_duration_since(self, earlier: Instant) -> Duration | None: ...
-    def checked_sub(self, other: Duration) -> Instant | None: ...
-    def duration_since(self, earlier: Instant) -> Duration: ...
-    def elapsed(self) -> Duration: ...
-    def saturating_duration_since(self, earlier: Instant) -> Duration: ...
-
-def instant() -> Instant: ...
-def sleep(seconds: float) -> float: ...
-
-# =============================================================================
-# STD::FS
-# =============================================================================
-class FileType:
-    def __repr__(self) -> str: ...
-    @property
-    def is_dir(self) -> bool: ...
-    @property
-    def is_file(self) -> bool: ...
-    @property
-    def is_symlink(self) -> bool: ...
-
-class Metadata:
-    def __repr__(self) -> str: ...
-    @property
-    def file_type(self) -> FileType: ...
-    @property
-    def len(self) -> int: ...
-    @property
-    def is_empty(self) -> bool: ...
-    @property
-    def modified(self) -> pydt.datetime: ...
-    @property
-    def accessed(self) -> pydt.datetime: ...
-    @property
-    def created(self) -> pydt.datetime: ...
-    @property
-    def is_dir(self) -> bool: ...
-    @property
-    def is_file(self) -> bool: ...
-    @property
-    def is_symlink(self) -> bool: ...
-
-# =============================================================================
-# STD::FS ~ functions
-# =============================================================================
-def read(path: FsPathLike) -> Bytes: ...
-def read_bytes(path: FsPathLike) -> bytes: ...
-def read_text(path: FsPathLike) -> str: ...
-def read_stream(
-    path: FsPathLike,
-    chunk_size: int = 65536,
-    *,
-    offset: int = 0,  # noqa: F811
-) -> t.Iterator[Bytes]: ...
-def write(path: FsPathLike, data: Buffer | str) -> int: ...
-def write_bytes(path: FsPathLike, data: bytes) -> int: ...
-def write_text(path: FsPathLike, data: str) -> int: ...
-def canonicalize(path: FsPathLike) -> FsPath: ...
-def copy(from_path: FsPathLike, to_path: FsPathLike) -> int: ...
-def create_dir(path: FsPathLike) -> None: ...
-def create_dir_all(path: FsPathLike) -> None: ...
-def exists(path: FsPathLike) -> bool: ...
-def is_dir(path: FsPathLike) -> bool: ...
-def is_file(path: FsPathLike) -> bool: ...
-def is_symlink(path: FsPathLike) -> bool: ...
-def metadata(path: FsPathLike) -> Metadata: ...
-def remove_dir(path: FsPathLike) -> None: ...
-def remove_dir_all(path: FsPathLike) -> None: ...
-def remove_file(path: FsPathLike) -> None: ...
-def rename(from_path: FsPathLike, to_path: FsPathLike) -> None: ...
-
-# =============================================================================
-# FSPATH
-# =============================================================================
-class FsPath:
-    def __init__(self, path: PathLike[str] | str | None = None) -> None: ...
-    def __fspath__(self) -> str: ...
-    def __str__(self) -> str: ...
-    def __repr__(self) -> str: ...
-    def __hash__(self) -> int: ...
-    def __eq__(self, other: object) -> bool: ...
-    def __ne__(self, other: object) -> bool: ...
-    def __lt__(self, other: PathLike[str] | str) -> bool: ...
-    def __le__(self, other: PathLike[str] | str) -> bool: ...
-    def __gt__(self, other: PathLike[str] | str) -> bool: ...
-    def __ge__(self, other: PathLike[str] | str) -> bool: ...
-    def __truediv__(self, other: PathLike[str] | str) -> FsPath: ...
-    def __rtruediv__(self, other: PathLike[str] | str) -> FsPath: ...
-    def to_pathlib(self) -> Path: ...
-    def read(self) -> Bytes: ...
-    def read_text(self) -> str: ...
-    def read_bytes(self) -> bytes: ...
-    def absolute(self) -> FsPath: ...
-    def resolve(self) -> FsPath: ...
-    def write(self, data: Buffer | bytes) -> None: ...
-    def write_bytes(self, data: Buffer | bytes) -> None: ...
-    def write_text(self, data: str) -> None: ...
-    def joinpath(self, *paths: str) -> FsPath: ...
-    def exists(self) -> bool: ...
-    def with_name(self, name: str) -> FsPath: ...
-    def with_suffix(self, suffix: str) -> FsPath: ...
-    def iterdir(self) -> t.Iterator[FsPath]: ...
-    def relative_to(self, other: PathLike[str] | str | FsPath) -> FsPath: ...
-    def as_posix(self) -> str: ...
-
-    # TODO
-    def __bytes__(self) -> bytes: ...
-    def as_uri(self) -> str: ...
-    def equiv(self, other: PathLike[str] | str | FsPath) -> bool: ...
-    def string(self) -> str: ...
-    def clone(self) -> FsPath: ...
-
-    # =========================================================================
-    # CLASSMETHODS
-    # =========================================================================
-    @classmethod
-    def cwd(cls) -> FsPath: ...
-    @classmethod
-    def home(cls) -> FsPath: ...
-
-    # =========================================================================
-    # PROPERTIES
-    # =========================================================================
-    @property
-    def anchor(self) -> str: ...
-    @property
-    def drive(self) -> str: ...
-    @property
-    def name(self) -> str: ...
-    @property
-    def parent(self) -> FsPath: ...
-    @property
-    def parents(self) -> t.Sequence[FsPath]: ...
-    @property
-    def parts(self) -> tuple[str, ...]: ...
-    @property
-    def root(self) -> str: ...
-    @property
-    def stem(self) -> str: ...
-    @property
-    def suffix(self) -> str: ...
-    @property
-    def suffixes(self) -> list[str]: ...
-
-    # =========================================================================
-    # std::path::PathBuf (deref -> std::path::Path)
-    # =========================================================================
-    def ancestors(self) -> t.Iterator[FsPath]: ...
-    def canonicalize(self) -> FsPath: ...
-    def components(self) -> t.Iterator[FsPath]: ...
-    def display(self) -> str: ...
-    def ends_with(self, path: PathLike[str] | str) -> bool: ...
-    def extension(self) -> str: ...
-    def file_name(self) -> str: ...
-    def file_prefix(self) -> FsPath: ...
-    def file_stem(self) -> str: ...
-    def has_root(self) -> bool: ...
-    def is_absolute(self) -> bool: ...
-    def is_dir(self) -> bool: ...
-    def is_file(self) -> bool: ...
-    def is_relative(self) -> bool: ...
-    def is_symlink(self) -> bool: ...
-    def starts_with(self, path: PathLike[str] | str) -> bool: ...
-    def strip_prefix(self, prefix: PathLike[str] | str) -> FsPath: ...
-    def with_extension(self, ext: str) -> FsPath: ...
-    def with_file_name(self, name: str) -> FsPath: ...
-
-FsPathLike = str | PathLike[str]
-
 def pwd() -> str: ...
 def home() -> str: ...
 def cd(path: FsPathLike) -> None: ...
@@ -437,32 +172,3 @@ def ls(
     objects: t.Literal[True] = True,
 ) -> list[FsPath]:
     """List directory contents - returns list of FsPath objects"""
-
-def quick_maths() -> t.Literal[3]:
-    """Performs quick-maths
-
-    Implements the algorithm for performing "quick-maths" as described by
-    Big Shaq in his PHD thesis, 2017, in which he states:
-
-    > "2 plus 2 is 4, minus one that's 3, quick maths." (Big Shaq et al., 2017)
-
-    Reference:
-        https://youtu.be/3M_5oYU-IsU?t=60
-
-    Example:
-        >>> import ry
-        >>> result = ry.quick_maths()
-        >>> assert result == 3
-
-    NOTE: THIS IS FROM MY TEMPLATE RY03-MODULE
-    """
-
-# -----------------------------------------------------------------------------
-# \/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\
-# /\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/
-# -----------------------------------------------------------------------------
-# |~|~ LIBS ~|~ LIBS ~|~ LIBS ~|~ LIBS ~|~ LIBS ~|~ LIBS ~|~ LIBS ~|~ LIBS ~|~|
-# -----------------------------------------------------------------------------
-# \/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\
-# /\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/\/
-# -----------------------------------------------------------------------------
