+++ python/ry/ryo3/_fspath.pyi
@@ -1,0 +1,104 @@
+"""ryo3-fspath types"""
+
+from __future__ import annotations
+
+import typing as t
+from os import PathLike
+from pathlib import Path
+
+from ry import Bytes
+from ry._types import Buffer
+
+# =============================================================================
+# FSPATH
+# =============================================================================
+class FsPath:
+    def __init__(self, path: PathLike[str] | str | None = None) -> None: ...
+    def __fspath__(self) -> str: ...
+    def __str__(self) -> str: ...
+    def __repr__(self) -> str: ...
+    def __hash__(self) -> int: ...
+    def __eq__(self, other: object) -> bool: ...
+    def __ne__(self, other: object) -> bool: ...
+    def __lt__(self, other: PathLike[str] | str) -> bool: ...
+    def __le__(self, other: PathLike[str] | str) -> bool: ...
+    def __gt__(self, other: PathLike[str] | str) -> bool: ...
+    def __ge__(self, other: PathLike[str] | str) -> bool: ...
+    def __truediv__(self, other: PathLike[str] | str) -> FsPath: ...
+    def __rtruediv__(self, other: PathLike[str] | str) -> FsPath: ...
+    def __bytes__(self) -> bytes: ...
+    def to_pathlib(self) -> Path: ...
+    def read(self) -> Bytes: ...
+    def read_text(self) -> str: ...
+    def read_bytes(self) -> bytes: ...
+    def absolute(self) -> FsPath: ...
+    def resolve(self) -> FsPath: ...
+    def write(self, data: Buffer | bytes) -> None: ...
+    def write_bytes(self, data: Buffer | bytes) -> None: ...
+    def write_text(self, data: str) -> None: ...
+    def joinpath(self, *paths: str) -> FsPath: ...
+    def exists(self) -> bool: ...
+    def with_name(self, name: str) -> FsPath: ...
+    def with_suffix(self, suffix: str) -> FsPath: ...
+    def iterdir(self) -> t.Iterator[FsPath]: ...
+    def relative_to(self, other: PathLike[str] | str | FsPath) -> FsPath: ...
+    def as_posix(self) -> str: ...
+    def as_uri(self) -> str: ...
+    def equiv(self, other: PathLike[str] | str | FsPath) -> bool: ...
+    def string(self) -> str: ...
+    def clone(self) -> FsPath: ...
+
+    # =========================================================================
+    # CLASSMETHODS
+    # =========================================================================
+    @classmethod
+    def cwd(cls) -> FsPath: ...
+    @classmethod
+    def home(cls) -> FsPath: ...
+
+    # =========================================================================
+    # PROPERTIES
+    # =========================================================================
+    @property
+    def anchor(self) -> str: ...
+    @property
+    def drive(self) -> str: ...
+    @property
+    def name(self) -> str: ...
+    @property
+    def parent(self) -> FsPath: ...
+    @property
+    def parents(self) -> t.Sequence[FsPath]: ...
+    @property
+    def parts(self) -> tuple[str, ...]: ...
+    @property
+    def root(self) -> str: ...
+    @property
+    def stem(self) -> str: ...
+    @property
+    def suffix(self) -> str: ...
+    @property
+    def suffixes(self) -> list[str]: ...
+
+    # =========================================================================
+    # std::path::PathBuf (deref -> std::path::Path)
+    # =========================================================================
+    def ancestors(self) -> t.Iterator[FsPath]: ...
+    def canonicalize(self) -> FsPath: ...
+    def components(self) -> t.Iterator[FsPath]: ...
+    def display(self) -> str: ...
+    def ends_with(self, path: PathLike[str] | str) -> bool: ...
+    def extension(self) -> str: ...
+    def file_name(self) -> str: ...
+    def file_prefix(self) -> FsPath: ...
+    def file_stem(self) -> str: ...
+    def has_root(self) -> bool: ...
+    def is_absolute(self) -> bool: ...
+    def is_dir(self) -> bool: ...
+    def is_file(self) -> bool: ...
+    def is_relative(self) -> bool: ...
+    def is_symlink(self) -> bool: ...
+    def starts_with(self, path: PathLike[str] | str) -> bool: ...
+    def strip_prefix(self, prefix: PathLike[str] | str) -> FsPath: ...
+    def with_extension(self, ext: str) -> FsPath: ...
+    def with_file_name(self, name: str) -> FsPath: ...
